% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutils.R
\name{catln}
\alias{catln}
\title{catln -- cat with newline at end (analogous to Pascal \code{writeln()})}
\usage{
catln(...)
}
\arguments{
\item{...}{(1)  -- arguments to pass to cat}
}
\value{
N/A
}
\description{
catln -- cat with newline at end (analogous to Pascal \code{writeln()})
}
\examples{
x=1; y=2; cc='banana'
catln(x,y,cc)

}
