% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutils.R
\name{genSampleDataFrameWithFactors}
\alias{genSampleDataFrameWithFactors}
\title{Generate a dummy dataframe with factors}
\usage{
genSampleDataFrameWithFactors()
}
\value{
Returns a test data frame
}
\description{
Generate a dummy dataframe with factors
}
\examples{
tdf <- genSampleDataFrameWithFactors()
print(summary(tdf))

}
