% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutils.R
\name{isQuote}
\alias{isQuote}
\title{Determine if character is a quote ('"`)}
\usage{
isQuote(ch)
}
\arguments{
\item{ch}{(1)  -- character to test}
}
\value{
--- boolean indicating whether ch is a quote char.
}
\description{
Determine if character is a quote ('"`)
}
