% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutils-package.R
\docType{package}
\name{nutils-package}
\alias{nutils}
\alias{nutils-package}
\title{nutils: Utilities for Debugging Mostly}
\description{
Provides simple utilities for early debugging and one-off scripts.
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
