% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutilsTestInteractive.R
\name{nutilsTestInteractive}
\alias{nutilsTestInteractive}
\title{nutilsTestInteractive.R
 Test \{nutils\} package interactive functions by hand.}
\usage{
nutilsTestInteractive()
}
\value{
Nothing. User decides if each test has worked.
}
\description{
nutilsTestInteractive.R
 Test \{nutils\} package interactive functions by hand.
}
\examples{
nutilsTestInteractive()
}
