% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutils.R
\name{parsetext}
\alias{parsetext}
\title{Abrreviates parse(text= paste0(.....)) useful for eval[.parent](parsetext(...)) ----
parsetext parse a character vector}
\usage{
parsetext(..., shouldShowText = FALSE)
}
\arguments{
\item{...}{-- text args to parse after pasting}

\item{shouldShowText}{-- boolean showText or not}
}
\value{
An object of type "expression" see ?parse
}
\description{
Abrreviates parse(text= paste0(.....)) useful for eval[.parent](parsetext(...)) ----
parsetext parse a character vector
}
\examples{
 parsetext('print("This is a test")')

}
