% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutils.R
\name{readStringFromClip}
\alias{readStringFromClip}
\title{ReadStringFromClip}
\usage{
readStringFromClip()
}
\arguments{
\item{None}{}
}
\value{
-- a string
}
\description{
Read a string (max 5000 char in mac) from clipboard
}
\details{
Uses writeChar on mac. See e.g. http://stackoverflow.com/a/13445458/1795127
}
