% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutils.R
\name{showvars}
\alias{showvars}
\title{Displays names and values of atomic variables (for texty debugging)}
\usage{
showvars(...)
}
\arguments{
\item{...}{(1) -- list of variables}
}
\value{
N/A (lists variables to console)
}
\description{
Displays names and values of atomic variables (for texty debugging)
}
