% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutilsTestInteractive.R
\name{testingTestInteractive}
\alias{testingTestInteractive}
\title{testingTestInteractive
 Test {testing} package interactive functions by hand}
\usage{
testingTestInteractive()
}
\value{
Nothing. User decides if each test has worked
}
\description{
testingTestInteractive
 Test {testing} package interactive functions by hand
}
\examples{
testingTestInteractive()
}
