% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutils.R
\name{writeStringToClip}
\alias{writeStringToClip}
\title{Write the string in inStr to clipboard}
\usage{
writeStringToClip(inStr)
}
\arguments{
\item{inStr}{string to copy to clipboard (uses writeChar on mac)}
}
\description{
Write the string in inStr to clipboard
}
