% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutils.R
\name{macedit}
\alias{macedit}
\title{Open a file in Mac OS text editor}
\usage{
macedit(fname)
}
\arguments{
\item{fname}{full path name of file to open}
}
\value{
N/A .. (Opens file in Text Edit app.)
}
\description{
Open a file in Mac OS text editor
}
\examples{
# not run
txtFile=system.file("extdata","myTextFile.txt",package='nutils')
macedit(txtFile)
}
