% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutils.R
\name{macopen}
\alias{macopen}
\title{Open directory in Mac OS finder, by default current working directory}
\usage{
macopen(dirOrFileName = getwd())
}
\arguments{
\item{dirOrFileName}{(1) = getwd()   path to folder to open in mac os}
}
\value{
NONE : If argument is a directory
}
\description{
Open directory in Mac OS finder, by default current working directory
}
