% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutils.R
\name{readStringFromClip}
\alias{readStringFromClip}
\title{Read a string (max 5000 char in mac) from clipboard}
\usage{
readStringFromClip(testText = NULL)
}
\arguments{
\item{testText}{a string/character array usually missing. It allows test text to be put in
and mostly gets around a roxygen2 issue I couldn't figure.}
}
\value{
a string
}
\description{
Read a string (max 5000 char in mac) from clipboard
}
\details{
Uses \code{writeChar()} on mac. See e.g. \url{http://stackoverflow.com/a/13445458/1795127}
}
\examples{
# Select someting and copy it to clipboard
# Then input a line
readline('Type cr when ready')
print(readStringFromClip())

}
