% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutils.R
\name{showvars}
\alias{showvars}
\title{Display names, types and values of variables (for quick wordy debugging)}
\usage{
showvars(...)
}
\arguments{
\item{...}{(1) -- list of variables}
}
\value{
N/A (lists variables to console)
}
\description{
Display names, types and values of variables (for quick wordy debugging)
}
\examples{
 x=1; y=2; cc='banana'
showvars(x,y,cc)

}
